requirements.txt
Django
django-rest-auth
djangorestframework
drf-generators
mysqlclient
